Restructured and documented src files for easier maintenance
