+++ src/tokenTypes.ts
@@ -1,0 +1,12 @@
+const tokenCodes = new Map<string, number>();
+tokenCodes.set("keyword", 0);
+
+const tokenTypesLegend = ["keyword"];
+
+const tokenModifiersLegend: string[] = [];
+
+export {
+    tokenCodes,
+    tokenTypesLegend,
+    tokenModifiersLegend
+}
